+++ moose/branches/msg_2008/utility/randnum/BinomialRng.h
@@ -28,7 +28,7 @@ class BinomialRng: public RandGenerator
 {
   public:
     BinomialRng();
-    void innerSetN(int n);
+    void innerSetN(unsigned long n);
     int innerGetN();
     void innerSetP(double p);
     double innerGetP();
